Move strip_accents to utils.py
