+++ openlibrary/plugins/ol_infobase.py
@@ -10,13 +10,13 @@ import os
 import re
 import sys
 import traceback
-import unicodedata
 
 import requests
 import six
 import web
 
 from infogami.infobase import cache, common, config, dbstore, server
+from openlibrary.plugins.upstream.utils import strip_accents
 
 from ..utils.isbn import isbn_10_to_isbn_13, isbn_13_to_isbn_10, normalize_isbn
 
@@ -591,14 +591,6 @@ class OLIndexer(_Indexer):
 
         if not isinstance(title, str):
             return ""
-
-        # http://stackoverflow.com/questions/517923/what-is-the-best-way-to-remove-accents-in-a-python-unicode-string
-        def strip_accents(s):
-            return ''.join(
-                c
-                for c in unicodedata.normalize('NFD', s)
-                if unicodedata.category(c) != 'Mn'
-            )
 
         norm = strip_accents(title).lower()
         norm = norm.replace(' and ', ' ')
